pyproject.toml
[project]
name = "openaiagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "fastapi-mcp>=0.1.4",
    "openai-agents>=0.0.8",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.4",
    "uvicorn>=0.34.0",
    "linkedin-api>=2.3.1",
]
